working but needs review
please take a look. It still needs to->The control flow class is complete. Now the user logs in, and is able to go the the dashboard. I have not implemented an add function to add a new user. -> if someone want to do this, awesome.
once at the dashboard, the user can select to add new households, all of this functionality works from start to finish. The only thing I noticed is that Economics class doesn't quite feel 'clean'. This can be adjusted. -> if someone want to do this, awesome.
The ability to read, write, and display the census data works. it's just not pretty.
I think it would be of everyone's best interest to run the code, respond back what you like or think needs improvement.